fixes and cleaning around thumbnail creation

+++ server/controllers/file.coffee
@@ -16,15 +16,19 @@ module.exports.fetch = (req, res, next, id) ->
         req.file = file
         next()
 
+# Return a list of file for a given month
 module.exports.list = (req, res, next) ->
+
     if req.params.page?
         skip = parseInt(req.params.page) * fileByPage
     else
         skip = 0
+
     [onCreation, percent] = onThumbCreation()
     if onCreation
-        res.send "percent": percent, 400
+        res.send "percent": percent
     else
+
         dates = {}
         options =
             limit: fileByPage + 1
@@ -34,11 +38,13 @@ module.exports.list = (req, res, next) ->
             if err
                 return res.error 500, 'An error occured', err
             else
+
                 if photos.length is fileByPage + 1
                     hasNext = true
                 else
                     hasNext = false
-                photos.splice(fileByPage,1)
+
+                photos.splice fileByPage, 1
                 for photo in photos
                     date = new Date(photo.lastModification)
                     mounth = date.getMonth() + 1
@@ -48,8 +54,11 @@ module.exports.list = (req, res, next) ->
                         dates[date].push photo
                     else
                         dates[date] = [photo]
-                res.send {files: dates, hasNext: hasNext}, 201
 
+                res.send {files: dates, hasNext: hasNext}, 200
+
+
+# Return thumb for given file.
 module.exports.thumb = (req, res, next) ->
     which = if req.file.binary.thumb then 'thumb' else 'file'
     stream = req.file.getBinary which, (err) ->
@@ -57,6 +66,7 @@ module.exports.thumb = (req, res, next) ->
     stream.pipe res
 
 
+# Resize picture
 resize = (raw, photo, name, callback) ->
     options = if name is 'thumb'
         mode: 'crop'
@@ -72,14 +82,17 @@ resize = (raw, photo, name, callback) ->
     options.dstPath = "/tmp/#{photo.id}2"
 
     # create files
-    fs.openSync options.dstPath, 'w'
+    fs.open options.dstPath, 'w', (err) ->
+        if err
+            callback err
+        else
+            # create a resized file and push it to db
+            im[options.mode] options, (err, stdout, stderr) ->
+                return callback err if err
+                photo.attachBinary options.dstPath, {name}, (err) ->
+                    fs.unlink options.dstPath, ->
+                        callback err
 
-    # create a resized file and push it to db
-    im[options.mode] options, (err, stdout, stderr) =>
-        return callback err if err
-        photo.attachBinary options.dstPath, {name}, (err) ->
-            fs.unlink options.dstPath, ->
-                callback err
 
 module.exports.createPhoto = (req, res, next) ->
     file = req.file
@@ -103,6 +116,7 @@ module.exports.createPhoto = (req, res, next) ->
             return next err if err
         stream.pipe fs.createWriteStream rawFile
         stream.on 'error', next
+
         stream.on 'end', =>
             if not photo.binary.thumb?
                 resize rawFile, photo, 'thumb', (err) ->
@@ -114,4 +128,3 @@ module.exports.createPhoto = (req, res, next) ->
                 resize rawFile, photo, 'screen', (err) ->
                     fs.unlink rawFile, ->
                         res.send 201, photo
-
